RDF fixes for ORNG and Authorhsip

+++ Database/ORNG/ORNG_CreateSchema.sql
@@ -584,18 +584,20 @@ BEGIN
 	IF (@SubjectID IS NULL)
 		SET @SubjectID = [RDF.].fnURI2NodeID(@SubjectURI)
 	
-	SELECT @InternalType = [Object] FROM [Ontology.Import].[Triple] 
+	SELECT @InternalType = REPLACE([Object], ' ', '') FROM [Ontology.Import].[Triple] 
 		WHERE [Subject] = 'http://orng.info/ontology/orng#ApplicationInstance' AND [Predicate] = 'http://www.w3.org/2000/01/rdf-schema#label'
-		
+	
 	SELECT @PersonID = cast(InternalID as INT), @InternalID = InternalID + '-' + CAST(@AppID as varchar) FROM [RDF.Stage].[InternalNodeMap]
 		WHERE [NodeID] = @SubjectID AND Class = 'http://xmlns.com/foaf/0.1/Person'
+
+	IF (@InternalID is null)
+	BEGIN
+		SELECT @InternalID = InternalID + '-GROUP-' + CAST(@AppID as varchar) FROM [RDF.Stage].[InternalNodeMap]
+			WHERE [NodeID] = @SubjectID AND Class = 'http://xmlns.com/foaf/0.1/Group'
+	END
 		
 	SELECT @PersonName = DisplayName from [Profile.Data].Person WHERE PersonID = @PersonID
-	--- this odd label format is required for the DataMap items to work properly!
-	SELECT @Label = 'http://orng.info/ontology/orng#ApplicationInstance^^' +
-					@InternalType + '^^' + @InternalID
-					
-					
+										
 	-- Convert the AppID to an AppName based on its URL
 	SELECT @AppName = REPLACE(RTRIM(RIGHT(url, CHARINDEX('/', REVERSE(url)) - 1)), '.xml', '')
 		FROM [ORNG.].[Apps] 
@@ -606,7 +608,7 @@ BEGIN
 				
 	SELECT @ApplicationNodeID  = NodeID
 		FROM [RDF.Stage].[InternalNodeMap]
-		WHERE Class = 'http://orng.info/ontology/orng#Application' AND InternalType = 'ORNG Application'
+		WHERE Class = 'http://orng.info/ontology/orng#Application' AND InternalType = 'ORNGApplication'
 			AND InternalID = @AppName
 
 		
@@ -616,7 +618,9 @@ BEGIN
 	DECLARE @AppInstanceID BIGINT
 	SELECT @AppInstanceID = NodeID
 		FROM [RDF.Stage].[InternalNodeMap]
-		WHERE Class = 'http://orng.info/ontology/orng#ApplicationInstance' AND InternalType = 'ORNG Application Instance'
+
+		WHERE Class = 'http://orng.info/ontology/orng#ApplicationInstance' 
+			AND InternalType = 'ORNGApplicationInstance'
 			AND InternalID = @InternalID
 	IF @AppInstanceID IS NOT NULL
 	BEGIN
@@ -637,73 +641,96 @@ BEGIN
 									 @ViewSecurityGroup = @ViewSecurityGroup, -- bigint
 									 @SessionID = NULL, -- uniqueidentifier
 									 @Error = NULL -- bit
-	END
-	ELSE
-	BEGIN
-		----------------------------------------------------------------
-		-- Add the app to the person for the first time
-		----------------------------------------------------------------
-		SELECT @Error = 0
-		BEGIN TRAN
-			-- We want Type 2.  Lookup internal type from import.triple, pass in AppID
-			EXEC [RDF.].GetStoreNode	@Class = 'http://orng.info/ontology/orng#ApplicationInstance',
-										@InternalType = @InternalType,
-										@InternalID = @InternalID,
-										@SessionID = @SessionID, 
-										@Error = @Error OUTPUT, 
-										@NodeID = @NodeID OUTPUT
-			-- for some reason, this Status in [RDF.Stage].InternalNodeMap is set to 0, not 3.  This causes issues so
-			-- we fix
-			UPDATE [RDF.Stage].[InternalNodeMap] SET [Status] = 3 WHERE NodeID = @NodeID						
-			
-			EXEC [RDF.].GetStoreNode @Value = @Label, @Language = NULL, @DataType = NULL,
-				@SessionID = @SessionID, @Error = @Error OUTPUT, @NodeID = @LabelID OUTPUT	
 
-			-- Add the Type
-			EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
-										@PredicateURI = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
-										@ObjectURI = 'http://orng.info/ontology/orng#ApplicationInstance',
-										@SessionID = @SessionID,
-										@Error = @Error OUTPUT
-			-- Add the Label
-			EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
-										@PredicateURI = 'http://www.w3.org/2000/01/rdf-schema#label',
-										@ObjectID = @LabelID,
-										@SessionID = @SessionID,
-										@Error = @Error OUTPUT
-			-- Link the ApplicationInstance to the Application
-			EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
-										@PredicateURI = 'http://orng.info/ontology/orng#applicationInstanceOfApplication',
-										@ObjectID = @ApplicationNodeID,
-										@SessionID = @SessionID,
-										@Error = @Error OUTPUT		
-			-- Link the ApplicationInstance to the person
-			EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
-										@PredicateURI = 'http://orng.info/ontology/orng#applicationInstanceForPerson',
-										@ObjectID = @SubjectID,
-										@SessionID = @SessionID,
-										@Error = @Error OUTPUT								
-			-- Link the person to the ApplicationInstance
-			EXEC [RDF.].GetStoreTriple	@SubjectID = @SubjectID,
-										@PredicateURI = @PredicateURI,
-										@ObjectID = @NodeID,
-										@SessionID = @SessionID,
-										@Error = @Error OUTPUT
-		COMMIT	
+		print 'Weare ready to add the person to the filter'
+		SELECT @PERSON_FILTER_ID = (SELECT PersonFilterID FROM [ORNG.].[Apps]  
+			WHERE AppID = @AppID AND PersonFilterID NOT IN (
+				SELECT personFilterId FROM [Profile.Data].[Person.FilterRelationship] 
+					WHERE PersonID = @PersonID))
+		IF (@PERSON_FILTER_ID IS NOT NULL) 
+			BEGIN
+				INSERT [Profile.Import].[PersonFilterFlag]
+					SELECT InternalUserName, PersonFilter FROM [Profile.Data].[Person], [Profile.Data].[Person.Filter]
+						WHERE PersonID = @PersonID AND PersonFilterID = @PERSON_FILTER_ID
+				INSERT [Profile.Data].[Person.FilterRelationship](PersonID, personFilterId) 
+					values (@PersonID, @PERSON_FILTER_ID)
+			END
+
+		-- Exit the proc
+		RETURN;
 	END
+
+
+	----------------------------------------------------------------
+	-- Add the app to the person for the first time
+	----------------------------------------------------------------
+	SELECT @Error = 0
+	BEGIN TRAN
+		-- We want Type 2.  Lookup internal type from import.triple, pass in AppID
+		EXEC [RDF.].GetStoreNode	@Class = 'http://orng.info/ontology/orng#ApplicationInstance',
+									@InternalType = @InternalType,
+									@InternalID = @InternalID,
+									@SessionID = @SessionID, 
+									@Error = @Error OUTPUT, 
+									@NodeID = @NodeID OUTPUT
+		SELECT @Label = 'ORNGApplicationInstance ' + cast(@NodeID as varchar)
+
+		-- for some reason, this Status in [RDF.Stage].InternalNodeMap is set to 0, not 3.  This causes issues so
+		-- we fix
+		--UPDATE [RDF.Stage].[InternalNodeMap] SET [Status] = 3 WHERE NodeID = @NodeID						
 			
-	-- wire in the filter to both the import and live tables
-	SELECT @PERSON_FILTER_ID = (SELECT PersonFilterID FROM Apps WHERE AppID = @AppID AND PersonFilterID NOT IN (
-			SELECT personFilterId FROM [Profile.Data].[Person.FilterRelationship] WHERE PersonID = @PersonID))
-	IF (@PERSON_FILTER_ID IS NOT NULL) 
-		BEGIN
-			INSERT [Profile.Import].[PersonFilterFlag]
-				SELECT InternalUserName, PersonFilter FROM [Profile.Data].[Person], [Profile.Data].[Person.Filter]
-					WHERE PersonID = @PersonID AND PersonFilterID = @PERSON_FILTER_ID
-			INSERT [Profile.Data].[Person.FilterRelationship](PersonID, personFilterId) values (@PersonID, @PERSON_FILTER_ID)
-		END
+		EXEC [RDF.].GetStoreNode @Value = @Label, @Language = NULL, @DataType = NULL,
+			@SessionID = @SessionID, @Error = @Error OUTPUT, @NodeID = @LabelID OUTPUT	
+
+		-- Add the Type
+		EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
+									@PredicateURI = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#type',
+									@ObjectURI = 'http://orng.info/ontology/orng#ApplicationInstance',
+									@SessionID = @SessionID,
+									@Error = @Error OUTPUT
+		-- Add the Label
+		EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
+									@PredicateURI = 'http://www.w3.org/2000/01/rdf-schema#label',
+									@ObjectID = @LabelID,
+									@SessionID = @SessionID,
+									@Error = @Error OUTPUT
+		-- Link the ApplicationInstance to the Application
+		EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
+									@PredicateURI = 'http://orng.info/ontology/orng#applicationInstanceOfApplication',
+									@ObjectID = @ApplicationNodeID,
+									@SessionID = @SessionID,
+									@Error = @Error OUTPUT		
+		-- Link the ApplicationInstance to the person
+		EXEC [RDF.].GetStoreTriple	@SubjectID = @NodeID,
+									@PredicateURI = 'http://orng.info/ontology/orng#applicationInstanceForPerson',
+									@ObjectID = @SubjectID,
+									@SessionID = @SessionID,
+									@Error = @Error OUTPUT								
+		-- Link the person to the ApplicationInstance
+		EXEC [RDF.].GetStoreTriple	@SubjectID = @SubjectID,
+									@PredicateURI = @PredicateURI,
+									@ObjectID = @NodeID,
+									@SessionID = @SessionID,
+									@Error = @Error OUTPUT
+		
+		-- wire in the filter to both the import and live tables
+		print 'We are ready to add the person to the filter'
+		SELECT @PERSON_FILTER_ID = (SELECT PersonFilterID FROM [ORNG.].[Apps]  
+
+			WHERE AppID = @AppID AND PersonFilterID NOT IN (
+				SELECT personFilterId FROM [Profile.Data].[Person.FilterRelationship] 
+					WHERE PersonID = @PersonID))
+		IF (@PERSON_FILTER_ID IS NOT NULL) 
+			BEGIN
+				INSERT [Profile.Import].[PersonFilterFlag]
+					SELECT InternalUserName, PersonFilter FROM [Profile.Data].[Person], [Profile.Data].[Person.Filter]
+						WHERE PersonID = @PersonID AND PersonFilterID = @PERSON_FILTER_ID
+
+				INSERT [Profile.Data].[Person.FilterRelationship](PersonID, personFilterId) 
+					values (@PersonID, @PERSON_FILTER_ID)
+			END
+	COMMIT	
 END
-
 GO
 
 /****** Object:  StoredProcedure [ORNG.].[GetAppInstance]    Script Date: 06/19/2014 09:37:50 ******/
@@ -752,33 +779,51 @@ SET QUOTED_IDENTIFIER ON
 GO
 
 CREATE PROCEDURE [ORNG.].[RemoveAppFromPerson]
-@SubjectID BIGINT=NULL, @SubjectURI nvarchar(255)=NULL, @AppID INT, @DeleteType tinyint = 1, @SessionID UNIQUEIDENTIFIER=NULL, @Error BIT=NULL OUTPUT
+@SubjectID BIGINT=NULL, @SubjectURI NVARCHAR(255)=NULL, @AppID INT, @DeleteType tinyint = 1, @SessionID UNIQUEIDENTIFIER=NULL, @Error BIT=NULL OUTPUT
 AS
 BEGIN
 	-- SET NOCOUNT ON added to prevent extra result sets from
 	-- interfering with SELECT statements.
 	SET NOCOUNT ON;
-	DECLARE @ApplicationInstanceNodeID BIGINT
-	DECLARE @PersonID INT	
+	DECLARE @ApplicationInstanceNodeID  BIGINT
 	DECLARE @TripleID BIGINT
+	DECLARE @PersonID INT
+	DECLARE @InternalID nvarchar(100)	
+	DECLARE @InternalType nvarchar(300)
 	DECLARE @PERSON_FILTER_ID INT
-	DECLARE @InternalUserName nvarchar(50)
-	DECLARE @PersonFilter nvarchar(50)
+	DECLARE @InternalUserName NVARCHAR(50)
+	DECLARE @PersonFilter NVARCHAR(50)
 
 	IF (@SubjectID IS NULL)
 		SET @SubjectID = [RDF.].fnURI2NodeID(@SubjectURI)
 	
-	EXEC [ORNG.].[GetAppInstance] @SubjectID = @SubjectID, 
-							 @AppID = @AppID, 
-							 @SessionID = @SessionID, 
-							 @ApplicationInstanceNodeID = @ApplicationInstanceNodeID OUTPUT,
-							 @Error = @Error OUTPUT							 
-							 
+	-- Lookup the PersonID
+	SELECT @InternalID = InternalID, @InternalType = InternalType
+		FROM [RDF.Stage].[InternalNodeMap]
+		WHERE NodeID = @SubjectID
+	
+	IF @InternalType = 'Person'
+	BEGIN
+		-- Lookup the App Instance's NodeID
+		SELECT @ApplicationInstanceNodeID  = NodeID
+			FROM [RDF.Stage].[InternalNodeMap]
+			WHERE Class = 'http://orng.info/ontology/orng#ApplicationInstance' AND InternalType = 'ORNGApplicationInstance'
+				AND InternalID = @InternalID + '-' + CAST(@AppID AS VARCHAR(50))
+	END
+	ELSE IF @InternalType = 'Group'
+	BEGIN
+		-- Lookup the App Instance's NodeID
+		SELECT @ApplicationInstanceNodeID  = NodeID
+			FROM [RDF.Stage].[InternalNodeMap]
+			WHERE Class = 'http://orng.info/ontology/orng#ApplicationInstance' AND InternalType = 'ORNGApplicationInstance'
+				AND InternalID = @InternalID + '-GROUP-' + CAST(@AppID AS VARCHAR(50))
+	END
+		
 	-- there is only ONE link from the person to the application object, so grab it	
 	SELECT @TripleID = [TripleID] FROM [RDF.].Triple 
 		WHERE [Subject] = @SubjectID
 		AND [Object] = @ApplicationInstanceNodeID
-		
+
 	-- now delete it
 	BEGIN TRAN
 
@@ -793,23 +838,24 @@ BEGIN
 							   @SessionID = @SessionID, 
 							   @Error = @Error OUTPUT
 		END							   
-							   
+
 		-- remove any filters
 		SELECT @PERSON_FILTER_ID = (SELECT PersonFilterID FROM Apps WHERE AppID = @AppID)
 		IF (@PERSON_FILTER_ID IS NOT NULL) 
 			BEGIN
-				SELECT @PersonID = cast(InternalID as INT) FROM [RDF.Stage].[InternalNodeMap]
+				SELECT @PersonID = CAST(InternalID AS INT) FROM [RDF.Stage].[InternalNodeMap]
 					WHERE [NodeID] = @SubjectID AND Class = 'http://xmlns.com/foaf/0.1/Person'
+				IF (@PersonID IS NOT NULL)
+					BEGIN
+						SELECT @InternalUserName = InternalUserName FROM [Profile.Data].[Person] WHERE PersonID = @PersonID
+						SELECT @PersonFilter = PersonFilter FROM [Profile.Data].[Person.Filter] WHERE PersonFilterID = @PERSON_FILTER_ID
 
-				SELECT @InternalUserName = InternalUserName FROM [Profile.Data].[Person] WHERE PersonID = @PersonID
-				SELECT @PersonFilter = PersonFilter FROM [Profile.Data].[Person.Filter] WHERE PersonFilterID = @PERSON_FILTER_ID
-
-				DELETE FROM [Profile.Import].[PersonFilterFlag] WHERE InternalUserName = @InternalUserName AND personfilter = @PersonFilter
-				DELETE FROM [Profile.Data].[Person.FilterRelationship] WHERE PersonID = @PersonID AND personFilterId = @PERSON_FILTER_ID
+						DELETE FROM [Profile.Import].[PersonFilterFlag] WHERE InternalUserName = @InternalUserName AND personfilter = @PersonFilter
+						DELETE FROM [Profile.Data].[Person.FilterRelationship] WHERE PersonID = @PersonID AND personFilterId = @PERSON_FILTER_ID
+					END
 			END
 	COMMIT
 END
-
 GO
 
 /****** Object:  StoredProcedure [ORNG.].[HasApp]    Script Date: 06/25/2014 11:36:04 ******/
